added max and min voltage, started on safety

+++ Core/Inc/LTC6813.h
@@ -359,7 +359,7 @@ void LTC6813_set_discharge(int Cell, //!< The cell to be discharged
  Helper Function to Set DCC bits in the CFGR Registers
  @return void
  */
-void LTC6813_set_discharge_per_segment(int Cell, //!< The cell to be discharged
+void LTC6813_set_discharge_per_segment(uint8_t Cell, //!< The cell to be discharged
                            uint8_t segment, //!< Number of ICs in the system
                            cell_asic *ic //!< A two dimensional array that will store the data
 						   );
