+++ Core/Src/LTC6813.c
@@ -441,7 +441,7 @@ void LTC6813_set_discharge(int Cell, // Cell to be discharged
 }
 
 /* Helper function to set discharge bit in CFG register */
-void LTC6813_set_discharge_per_segment(int Cell, // Cell to be discharged
+void LTC6813_set_discharge_per_segment(uint8_t Cell, // Cell to be discharged
 						   uint8_t segment, // Number of ICs in the system
 						   cell_asic *ic // A two dimensional array that will store the data
 						   )
